Ran into problems due to the delete failing when the checkexists task fails. Tried numerous ways of getting this to work and then discovered in the documentation that all batch jobs should check to see if the batch has been cancelled. By implementing this, and implementing a cancel of batch in the check exists job, we can check if the host exists, if it doesn't then the other jobs are run but their functionality bipassed, resulting in a host that gets deleted.
#603 5.1 - NSX | Dedicated Hosts - Delete host listener

+++ app/Jobs/Artisan/Host/RemoveFrom3Par.php
@@ -22,27 +22,23 @@ class RemoveFrom3Par extends Job
     public function handle()
     {
         Log::info(get_class($this) . ' : Started', ['id' => $this->model->id]);
+        if (!$this->batch()->cancelled()) {
+            $host = $this->model;
+            $availabilityZone = $host->hostGroup->availabilityZone;
 
-        $host = $this->model;
-        $availabilityZone = $host->hostGroup->availabilityZone;
-
-        try {
-            $availabilityZone->artisanService()->delete(
-                '/api/v2/san/' . $availabilityZone->san_name . '/host/' . $host->id
-            );
-        } catch (RequestException $exception) {
-            Log::error(get_class($this) . ' : Failed', [
-                'id' => $host->id,
-                'status_code' => $exception->getCode(),
-                'content' => $exception->getResponse()->getBody()->getContents()
-            ]);
-            if ($exception->getCode() != 404) {
-                throw $exception;
+            try {
+                $availabilityZone->artisanService()->delete(
+                    '/api/v2/san/' . $availabilityZone->san_name . '/host/' . $host->id
+                );
+            } catch (RequestException $exception) {
+                if ($exception->getCode() != 404) {
+                    $this->fail($exception);
+                    throw $exception;
+                }
+                Log::warning(get_class($this) . ' : Host ' . $host->id . ' was not removed from 3Par.');
+                return;
             }
-            Log::warning(get_class($this) . ' : Host ' . $host->id . ' was not removed from 3Par.');
-            return;
         }
-
         Log::info(get_class($this) . ' : Finished', ['id' => $this->model->id]);
     }
 }
