+++ app/Jobs/Conjurer/Host/PowerOff.php
@@ -22,28 +22,24 @@ class PowerOff extends Job
     public function handle()
     {
         Log::info(get_class($this) . ' : Started', ['id' => $this->model->id]);
+        if (!$this->batch()->cancelled()) {
+            $host = $this->model;
+            $hostGroup = $host->hostGroup;
+            $availabilityZone = $host->hostGroup->availabilityZone;
 
-        $host = $this->model;
-        $hostGroup = $host->hostGroup;
-        $availabilityZone = $host->hostGroup->availabilityZone;
-
-        try {
-            $availabilityZone->conjurerService()->delete(
-                '/api/v2/compute/' . $availabilityZone->ucs_compute_name . '/vpc/' . $hostGroup->vpc->id . '/host/' . $host->id . '/power'
-            );
-        } catch (RequestException $exception) {
-            Log::error(get_class($this) . ' : Failed', [
-                'id' => $host->id,
-                'status_code' => $exception->getCode(),
-                'content' => $exception->getResponse()->getBody()->getContents()
-            ]);
-            if ($exception->getCode() != 404) {
-                throw $exception;
+            try {
+                $availabilityZone->conjurerService()->delete(
+                    '/api/v2/compute/' . $availabilityZone->ucs_compute_name . '/vpc/' . $hostGroup->vpc->id . '/host/' . $host->id . '/power'
+                );
+            } catch (RequestException $exception) {
+                if ($exception->getCode() != 404) {
+                    $this->fail($exception);
+                    throw $exception;
+                }
+                Log::warning(get_class($this) . ' : Host ' . $host->id . ' was not powered off.');
+                return;
             }
-            Log::warning(get_class($this) . ' : Host ' . $host->id . ' was not powered off.');
-            return;
         }
-
         Log::info(get_class($this) . ' : Finished', ['id' => $this->model->id]);
     }
 }
