+++ app/Jobs/Kingpin/Host/DeleteInVmware.php
@@ -25,46 +25,46 @@ class DeleteInVmware extends Job
     public function handle()
     {
         Log::info(get_class($this) . ' : Started', ['id' => $this->host->id]);
+        if (!$this->batch()->cancelled()) {
+            $availabilityZone = $this->host->hostGroup->availabilityZone;
+            $hostGroup = $this->host->hostGroup;
 
-        $availabilityZone = $this->host->hostGroup->availabilityZone;
-        $hostGroup = $this->host->hostGroup;
-
-        // Get the host spec from Conjurer
-        try {
-            $response = $availabilityZone->conjurerService()->get(
-                '/api/v2/compute/' . $availabilityZone->ucs_compute_name . '/vpc/' . $hostGroup->vpc->id . '/host/' . $this->host->id
-            );
-            $response = json_decode($response->getBody()->getContents());
-        } catch (RequestException $exception) {
-            if ($exception->getCode() != 404) {
-                throw $exception;
+            // Get the host spec from Conjurer
+            try {
+                $response = $availabilityZone->conjurerService()->get(
+                    '/api/v2/compute/' . $availabilityZone->ucs_compute_name . '/vpc/' . $hostGroup->vpc->id . '/host/' . $this->host->id
+                );
+                $response = json_decode($response->getBody()->getContents());
+            } catch (RequestException $exception) {
+                if ($exception->getCode() != 404) {
+                    throw $exception;
+                }
+                Log::warning(get_class($this) . ' : Host was not found on UCS, skipping.');
+                return;
             }
-            Log::warning(get_class($this) . ' : Host was not found on UCS, skipping.');
-            return;
-        }
 
-        $macAddress = collect($response->interfaces)->firstWhere('name', '=', 'eth0')->address;
-        if (empty($macAddress)) {
-            $message = 'Failed to load eth0 address for host ' . $this->host->id;
-            Log::error($message);
-            $this->fail(new \Exception($message));
-            return false;
-        }
-
-        Log::debug('MAC address: ' . $macAddress);
-
-        try {
-            $availabilityZone->kingpinService()->delete(
-                '/api/v2/vpc/' . $hostGroup->vpc_id . '/hostgroup/' . $hostGroup->id . '/host/' . $macAddress
-            );
-        } catch (RequestException $exception) {
-            if ($exception->getCode() != 404) {
-                throw $exception;
+            $macAddress = collect($response->interfaces)->firstWhere('name', '=', 'eth0')->address;
+            if (empty($macAddress)) {
+                $message = 'Failed to load eth0 address for host ' . $this->host->id;
+                Log::error($message);
+                $this->fail(new \Exception($message));
+                return false;
             }
-            Log::warning(get_class($this) . ' : Host could not be deleted, skipping.');
-            return;
-        }
 
+            Log::debug('MAC address: ' . $macAddress);
+
+            try {
+                $availabilityZone->kingpinService()->delete(
+                    '/api/v2/vpc/' . $hostGroup->vpc_id . '/hostgroup/' . $hostGroup->id . '/host/' . $macAddress
+                );
+            } catch (RequestException $exception) {
+                if ($exception->getCode() != 404) {
+                    throw $exception;
+                }
+                Log::warning(get_class($this) . ' : Host could not be deleted, skipping.');
+                return;
+            }
+        }
         Log::info(get_class($this) . ' : Finished', ['id' => $this->host->id]);
     }
 }
